Fix a bug with generate_search_blocks
I think what's happening here is that generate_search_blocks
is doing the right thing, but only if the job never stops and restarts,
and only because the starting_from mutates the internal state
and that is remembered for the next check.

If the job restarts, it would lose that internal state and start over.

This also suggests that we should have a verification step
that asserts the search blocks aren't skipping anything,
perhaps that each block as a well-defined "successor" block
whose next starting value is +1 the last ending value,
or whose starting value is zero and the next level is +1 the last level.